+++ src/templates/default.js
@@ -2,10 +2,6 @@
 
 module.exports = {
 	subject: '[Rescuefy] - Hay un nuevo mensaje para vos!',
-	to: [
-		'rgpxen+rescuefy@gmail.com',
-		'proyecto.rescuefy+default@gmail.com'
-	],
 	// eslint-disable-next-line max-len
 	html: '<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "https://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd"><html xmlns="https://www.w3.org/1999/xhtml" style="height:100%;"><head><title>Rescuefy Default Email</title><meta http-equiv="Content-Type" content="text/html; charset=UTF-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width, initial-scale=1.0 "><style>*::after{box-sizing:border-box}*:before{box-sizing:border-box}.rescuefy-email__footer-rescuefy:hover{color:#6c2650 !important}</style></head><body class="rescuefy-email__body" style="box-sizing:border-box;font-family:sans-serif;text-align:center;background-color:#9ae2fe;height:100%;min-width:fit-content;margin:0;"><table class="rescuefy-email__container" style="box-sizing:border-box;border-spacing:0;margin:auto;height:675px;max-width:540px;width:90%;padding:0 10px;"><thead style="box-sizing:border-box;"><th class="rescuefy-email__header" style="box-sizing:border-box;padding:0px;margin:0;padding:86px 0 56px 0;"><img src="https://user-images.githubusercontent.com/89747340/132135868-755c19bf-4313-40cf-a997-1710705e6142.png" alt="logo-rescuefy-mails" style="box-sizing:border-box;width:100%;"></th></thead><tbody style="box-sizing:border-box;"><tr style="box-sizing:border-box;"><td class="rescuefy-email__message" style="box-sizing:border-box;padding:0px;margin:0;background-color:#cfffff;color:#5e4af9;min-height:347px;width:540px;border-radius:3px;padding:56px 10px 10px;"><h1 class="rescuefy-email__greetings" style="box-sizing:border-box;margin:0;padding:0 115px 35px 113px;font-size:30px;font-weight:500;line-height:40px;color:#5e4af9;">{{title}}</h1><p class="rescuefy-email__info-text" style="box-sizing:border-box;font-size:18px;color:#5e4af9;font-weight:400;width:312px;margin:0 auto;line-height:25px;">{{mainText}}</p><br style="box-sizing:border-box;"/><p class="rescuefy-email__info-text" style="box-sizing:border-box;font-size:18px;color:#5e4af9;font-weight:400;width:312px;margin:0 auto;line-height:25px;">{{secondaryText}}</p></td></tr></tbody><tfoot style="box-sizing:border-box;"><tr style="box-sizing:border-box;"><td class="rescuefy-email__footer" style="box-sizing:border-box;padding:0px;margin:0;padding:43px 0 28px 0;"><a href="https://github.com/Rescuefy" class="rescuefy-email__footer-rescuefy" style="box-sizing:border-box;font-size:18px;color:#5e4af9;font-weight:500;margin-bottom:20px;display:inline-block;text-decoration:none;">Rescuefy en Github</a><p class="rescuefy-email__footer-text" style="box-sizing:border-box;color:#6c2650;font-size:8px;line-height:13px;font-weight:400;margin:0;">Copyright &copy; 2021 Rescuefy</p><p class="rescuefy-email__footer-text" style="box-sizing:border-box;color:#6c2650;font-size:8px;line-height:13px;font-weight:400;margin:0;">Buenos Aires, Argentina</p><p class="rescuefy-email__footer-text" style="box-sizing:border-box;color:#6c2650;font-size:8px;line-height:13px;font-weight:400;margin:0;">No responda este email!</p></td></tr></tfoot></table><hr class="rescuefy-email__footer-line" style="box-sizing:border-box;height:4px;background-color:#5e4af9;border:none;margin:0;"></body></html>'
 };
